/sdk/basyx/object_store.py Rename objectstore.py to match python naming style
